@@ -12,9 +12,9 @@
   <artifactId>SimpleCodeTester-Executor</artifactId>
 
   <properties>
-    <maven.compiler.source>11</maven.compiler.source>
-    <maven.compiler.target>11</maven.compiler.target>
-    <asm.version>9.2</asm.version>
+    <maven.compiler.source>17</maven.compiler.source>
+    <maven.compiler.target>17</maven.compiler.target>
+    <asm.version>9.3</asm.version>
   </properties>
 
   <build>
