@@ -17,19 +17,19 @@
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <immutables.version>2.8.8</immutables.version>
-    <junit.version>5.8.2</junit.version>
+    <immutables.version>2.9.1</immutables.version>
+    <junit.version>5.9.0</junit.version>
     <slf4j.version>1.7.30</slf4j.version>
     <logback.version>1.2.10</logback.version>
-    <gson.version>2.8.9</gson.version>
-    <lombok.version>1.18.22</lombok.version>
+    <gson.version>2.9.0</gson.version>
+    <lombok.version>1.18.24</lombok.version>
     <hamcrest.version>2.2</hamcrest.version>
     <jsr305.version>3.0.2</jsr305.version>
 
     <maven-compiler-plugin.version>3.8.1</maven-compiler-plugin.version>
     <maven-shade-plugin.version>3.2.2</maven-shade-plugin.version>
     <maven-surefire-plugin.version>3.0.0-M4</maven-surefire-plugin.version>
-    <jacoco-maven-plugin.version>0.8.5</jacoco-maven-plugin.version>
+    <jacoco-maven-plugin.version>0.8.8</jacoco-maven-plugin.version>
     <maven-failsafe-plugin.version>3.0.0-M5</maven-failsafe-plugin.version>
     <versions-maven-plugin.version>2.7</versions-maven-plugin.version>
     <maven-enforcer-plugin.version>3.0.0-M3</maven-enforcer-plugin.version>
@@ -41,7 +41,7 @@
         <artifactId>maven-compiler-plugin</artifactId>
         <version>${maven-compiler-plugin.version}</version>
         <configuration>
-          <release>11</release>
+          <release>17</release>
         </configuration>
       </plugin>
 
