@@ -1,7 +1,7 @@
 package me.ialistannen.simplecodetester.backend.endpoints;
 
+import jakarta.validation.constraints.NotEmpty;
 import java.util.List;
-import javax.validation.constraints.NotEmpty;
 import lombok.extern.slf4j.Slf4j;
 import me.ialistannen.simplecodetester.backend.db.entities.CheckCategory;
 import me.ialistannen.simplecodetester.backend.services.checks.CheckCategoryService;
